requirements.txt
yt-dlp
requests
python-dotenv
num2words
google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
lxml
pydantic
filelock
fastapi
uvicorn[standard]
python-multipart
PyJWT
bcrypt
email-validator
